api/requirements.txt
Django==5.1.10
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg[binary]==3.2.9
whitenoise==6.7.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
dj-database-url==2.2.0
stripe==10.6.0
PyJWT==2.9.0
djangorestframework-simplejwt==5.5.1
reportlab==4.2.2
python-docx==1.1.2
ruff==0.6.9
pdfminer.six==20240706
PyYAML==6.0.2
requests==2.32.3
